+++ proc-macro-definitions/src/lib.rs
@@ -7,7 +7,7 @@ use lazy_static::lazy_static;
 use proc_macro::TokenStream as TokenStream1;
 use proc_macro2::{Span, TokenStream as TokenStream2};
 use proc_macro_crate::{crate_name, FoundCrate};
-use quote::{quote, quote_spanned, ToTokens};
+use quote::{quote, quote_spanned};
 use std::iter;
 use syn::{
 	parse::{Parse, ParseStream},
@@ -20,7 +20,6 @@ mod part;
 mod storage_configuration;
 mod storage_context;
 mod syn_ext;
-mod trace_instrumentation;
 mod try_parse;
 
 use self::{
@@ -138,19 +137,6 @@ pub fn fragment(input: TokenStream1) -> TokenStream1 {
 	.into()
 }
 
-/// Iff the `"backtrace"` feature is enabled, instruments a function to add a trace frame of the form "attr_param::function_name".
-/// This only works on functions that return `Result<_, Escalation>`.
-#[proc_macro_attribute]
-pub fn trace_escalations(attr: TokenStream1, item: TokenStream1) -> TokenStream1 {
-	if cfg!(feature = "backtrace") {
-		let mut gui_traced = parse_macro_input!(item as Tracing);
-		gui_traced.prefix = parse_macro_input!(attr as TokenStream2).into();
-		gui_traced.into_token_stream().into()
-	} else {
-		item
-	}
-}
-
 // TODO: Accept reexported asteracea module made available via `use`.
 lazy_static! {
 	static ref ASTERACEA_NAME: String = crate_name("asteracea")
@@ -171,7 +157,6 @@ mod workaround_module {
 		const CAN_CAPTURE: bool;
 	}
 }
-use trace_instrumentation::Tracing;
 use workaround_module::Configuration;
 
 fn warn(location: Span, message: &str) -> Result<()> {